no finish

+++ src/cython/header/particle.h
@@ -3,5 +3,5 @@ typedef struct {
     double pos[3];
     double vel[3];
     double E;
-    long long id;
+    int id;
 } Particle;
